trying to fix client server
